refactor: barrel import on new icons
